Added requirements.txt and updated main.py
Now, there is a fully functioning progress indicator that uses the yaspin module.

Previously, the script would check if the local data stored was already today's date. However, if the data was updated through the day and the source file was updated, the script would not download the newest version.

I have removed this for simplicity purposes.

I'm not sure if its the conventional method of writing a requirements file, but it serves to show what modules the user may need to install.